nodes: offload the mmap first
comfy core will pin the module tensors in-place in some options
combinations.

This means migrating a module tensors after the load process is
invalid. So do the mmap offload up front.

I dont fully understand this errata where it nukes the mmap as
its not great for performance, but this should get over the crasher
bug users are experience as of this writing.